TensND

Documentation for TensND.

TensND.ϵ

TensND.Basis

TensND.CanonicalBasis

LinearAlgebra.normalize

TensND.basis

TensND.fϵ

TensND.isorthogonal

TensND.metric

TensND.normal_basis

TensND.ϵ — Constant

ϵ[i,j,k]

Levi-Civita symbol ϵᵢⱼₖ=(i-j)(j-k)(k-i)/2

TensND.Basis — Type

Basis{dim, T<:Number}(v, var = :cov)
Basis(θ::Number, ϕ::Number, ψ::Number)

Basis built from a square matrix v where columns correspond either to

primal vectors ie eᵢ=v[:,i] if var = :cov as by default

dual vectors ie eⁱ=v[:,i] if var = :cont.

Basis can also be built from Euler angles (θ, ϕ, ψ)

The attributes of this object are

Basis.e: square matrix defining the primal basis eᵢ=e[:,i]

Basis.E: square matrix defining the dual basis eⁱ=E[:,i]

Basis.g: square matrix defining the covariant components of the metric tensor gᵢⱼ=eᵢ⋅eⱼ=g[i,j]

Basis.G: square matrix defining the contravariant components of the metric tensor gⁱʲ=eⁱ⋅eʲ=G[i,j]

Examples

julia> using LinearAlgebra, SymPy

julia> v = Sym[1 0 0; 0 1 0; 0 1 1] ; b = Basis(v)
Basis{3, Sym}(Sym[1 0 0; 0 1 0; 0 1 1], Sym[1 0 0; 0 1 -1; 0 0 1], Sym[1 0 0; 0 2 1; 0 1 1], Sym[1 0 0; 0 1 -1; 0 -1 2])

julia> θ, ϕ, ψ = symbols("θ, ϕ, ψ", real = true) ; b = Basis(θ, ϕ, ψ) ; display(b.e)
3×3 Tensor{2, 3, Sym, 9}:
 -sin(ψ)⋅sin(ϕ) + cos(θ)⋅cos(ψ)⋅cos(ϕ)  -sin(ψ)⋅cos(θ)⋅cos(ϕ) - sin(ϕ)⋅cos(ψ)  sin(θ)⋅cos(ϕ)
  sin(ψ)⋅cos(ϕ) + sin(ϕ)⋅cos(θ)⋅cos(ψ)  -sin(ψ)⋅sin(ϕ)⋅cos(θ) + cos(ψ)⋅cos(ϕ)  sin(θ)⋅sin(ϕ)
                        -sin(θ)⋅cos(ψ)                          sin(θ)⋅sin(ψ)         cos(θ)

TensND.CanonicalBasis — Type

CanonicalBasis{dim, T}()

Canonical basis of dimension dim (default: 3) and type T (default: Sym)

The attributes of this object are

Basis.e: identity matrix defining the primal basis e[i,j]=δᵢⱼ

Basis.E: identity matrix defining the dual basis g[i,j]=δᵢⱼ

Basis.g: identity matrix defining the covariant components of the metric tensor g[i,j]=δᵢⱼ

Basis.G: identity matrix defining the contravariant components of the metric tensor G[i,j]=δᵢⱼ

Examples

julia> using LinearAlgebra, SymPy

julia> b = CanonicalBasis()
CanonicalBasis{3, Sym}(Sym[1 0 0; 0 1 0; 0 0 1], Sym[1 0 0; 0 1 0; 0 0 1], Sym[1 0 0; 0 1 0; 0 0 1], Sym[1 0 0; 0 1 0; 0 0 1])

julia> b = CanonicalBasis{2, Float64}()
CanonicalBasis{2, Float64}([1.0 0.0; 0.0 1.0], [1.0 0.0; 0.0 1.0], [1.0 0.0; 0.0 1.0], [1.0 0.0; 0.0 1.0])

LinearAlgebra.normalize — Function

normalize(b::AbstractBasis, var = cov)

Builds a normalized basis from the input basis b by calling normal_basis

TensND.basis — Method

basis(b::AbstractBasis, var = :cov)

Returns the primal (if var = :cov) or primal (if var = :cont) basis

TensND.fϵ — Method

fϵ(T, i::Int, j::Int, k::Int)
fϵ(i::Int, j::Int, k::Int) = fϵ(Int, i::Int, j::Int, k::Int)

Function giving Levi-Civita symbol ϵᵢⱼₖ=(i-j)(j-k)(k-i)/2

TensND.isorthogonal — Method

isorthogonal(b::AbstractBasis)

Checks whether the basis b is orthogonal

TensND.metric — Method

metric(b::AbstractBasis, var = :cov)

Returns the covariant (if var = :cov) or contravariant (if var = :cont) metric matrix

TensND.normal_basis — Method

normal_basis(v::AbstractArray{T,2}, var = :cov) where {T}

Builds a basis after normalization of column vectors of input matrix v where columns define either

primal vectors ie eᵢ=v[:,i]/norm(v[:,i]) if var = :cov as by default

dual vector ie eⁱ=v[:,i]/norm(v[:,i]) if var = :cont.