TensND

Documentation for TensND.

TensND.ϵ

TensND.Basis

TensND.CanonicalBasis

TensND.RotatedBasis

TensND.Tensnd

LinearAlgebra.normalize

TensND.components

TensND.fϵ

TensND.isorthogonal

TensND.isorthonormal

TensND.metric

TensND.normal_basis

TensND.vecbasis

TensND.ϵ — Constant

ϵ[i,j,k]

Levi-Civita symbol ϵᵢⱼₖ=(i-j)(j-k)(k-i)/2

TensND.Basis — Type

Basis(v::AbstractArray{T,2}, ::Val{:cov})
Basis{dim, T<:Number}()
Basis(θ::T<:Number, ϕ::T<:Number, ψ::T<:Number)

Basis built from a square matrix v where columns correspond either to

primal vectors ie eᵢ=v[:,i] if var=:cov as by default

dual vectors ie eⁱ=v[:,i] if var=:cont.

Basis without any argument refers to the canonical basis (CanonicalBasis) in Rᵈⁱᵐ (by default dim=3 and T=Sym)

Basis can also be built from Euler angles (RotatedBasis) θ in 2D and (θ, ϕ, ψ) in 3D

The attributes of this object are

Basis.e: square matrix defining the primal basis eᵢ=e[:,i]

Basis.E: square matrix defining the dual basis eⁱ=E[:,i]

Basis.g: square matrix defining the covariant components of the metric tensor gᵢⱼ=eᵢ⋅eⱼ=g[i,j]

Basis.G: square matrix defining the contravariant components of the metric tensor gⁱʲ=eⁱ⋅eʲ=G[i,j]

Examples

julia> v = Sym[1 0 0; 0 1 0; 0 1 1] ; b = Basis(v)
3×3 Basis{3, Sym}:
 1  0  0
 0  1  0
 0  1  1

julia> v = Sym[1 0 0; 0 1 0; 0 1 1] ; b = Basis(v, :cont)
3×3 Basis{3, Sym}:
 1  0   0
 0  1  -1
 0  0   1

julia> θ, ϕ, ψ = symbols("θ, ϕ, ψ", real = true) ; b = Basis(θ, ϕ, ψ) ; display(b.e)
3×3 Tensor{2, 3, Sym, 9}:
 -sin(ψ)⋅sin(ϕ) + cos(θ)⋅cos(ψ)⋅cos(ϕ)  -sin(ψ)⋅cos(θ)⋅cos(ϕ) - sin(ϕ)⋅cos(ψ)  sin(θ)⋅cos(ϕ)
  sin(ψ)⋅cos(ϕ) + sin(ϕ)⋅cos(θ)⋅cos(ψ)  -sin(ψ)⋅sin(ϕ)⋅cos(θ) + cos(ψ)⋅cos(ϕ)  sin(θ)⋅sin(ϕ)
                        -sin(θ)⋅cos(ψ)                          sin(θ)⋅sin(ψ)         cos(θ)

TensND.CanonicalBasis — Type

CanonicalBasis{dim, T}

Canonical basis of dimension dim (default: 3) and type T (default: Sym)

The attributes of this object are

Basis.e: identity matrix defining the primal basis e[i,j]=δᵢⱼ

Basis.E: identity matrix defining the dual basis g[i,j]=δᵢⱼ

Basis.g: identity matrix defining the covariant components of the metric tensor g[i,j]=δᵢⱼ

Basis.G: identity matrix defining the contravariant components of the metric tensor G[i,j]=δᵢⱼ

Examples

julia> b = CanonicalBasis()
3×3 CanonicalBasis{3, Sym}:
 1  0  0
 0  1  0
 0  0  1

julia> b = CanonicalBasis{2, Float64}()
2×2 CanonicalBasis{2, Float64}:
 1.0  0.0
 0.0  1.0

TensND.RotatedBasis — Type

RotatedBasis(θ::T<:Number, ϕ::T<:Number, ψ::T<:Number)
RotatedBasis(θ::T<:Number)

Orthonormal basis of dimension dim (default: 3) and type T (default: Sym) built from Euler angles θ in 2D and (θ, ϕ, ψ) in 3D

Examples

julia> θ, ϕ, ψ = symbols("θ, ϕ, ψ", real = true) ; b = RotatedBasis(θ, ϕ, ψ) ; display(b.e)
3×3 Tensor{2, 3, Sym, 9}:
 -sin(ψ)⋅sin(ϕ) + cos(θ)⋅cos(ψ)⋅cos(ϕ)  -sin(ψ)⋅cos(θ)⋅cos(ϕ) - sin(ϕ)⋅cos(ψ)  sin(θ)⋅cos(ϕ)
  sin(ψ)⋅cos(ϕ) + sin(ϕ)⋅cos(θ)⋅cos(ψ)  -sin(ψ)⋅sin(ϕ)⋅cos(θ) + cos(ψ)⋅cos(ϕ)  sin(θ)⋅sin(ϕ)
                        -sin(θ)⋅cos(ψ)                          sin(θ)⋅sin(ψ)         cos(θ)

TensND.Tensnd — Type

Tensnd{order,dim,TA<:Number,TB<:Number,A<:AbstractArray,B<:AbstractBasis}

Tensor type of any order defined by

a multiarray of components (of any type heriting from AbstractArray, e.g. Tensor or SymmetricTensor)

a basis of AbstractBasis type

a tuple of variances (covariant :cov or contravariant :cont) of length equal to the order of the tensor

Examples

julia> v = Sym[1 0 0; 0 1 0; 0 1 1] ; b = Basis(v)
3×3 Basis{3, Sym}:
 1  0  0
 0  1  0
 0  1  1

julia> T = Tensnd(b.g,(:cov,:cov),b)
3×3 Tensnd{2, 3, Sym, Sym, SymmetricTensor{2, 3, Sym, 6}, Basis{3, Sym}}:
 1  0  0
 0  2  1
 0  1  1

julia> components(T,(:cont,:cov),b)
3×3 Matrix{Sym}:
 1  0  0
 0  1  0
 0  0  1

LinearAlgebra.normalize — Function

normalize(b::AbstractBasis, var = cov)

Builds a normalized basis from the input basis b by calling normal_basis

TensND.components — Method

components(::Tensnd{order,dim,T}, ::NTuple{order,Symbol})
components(::Tensnd{order,dim,T}, ::NTuple{order,Symbol}, ::AbstractBasis{dim,T})

Extracts the components of a tensor for new variances and/or in a new basis

Examples

julia> v = Sym[0 1 1; 1 0 1; 1 1 0] ; b = Basis(v)
3×3 Basis{3, Sym}:
 0  1  1
 1  0  1
 1  1  0

julia> V = Tensor{1,3}(i->symbols("v$i",real=true))
3-element Vec{3, Sym}:
 v₁
 v₂
 v₃

julia> TV = Tensnd(V) # TV = Tensnd(V, (:cont,), CanonicalBasis())
3-element Tensnd{1, 3, Sym, Sym, Vec{3, Sym}, CanonicalBasis{3, Sym}}:
 v₁
 v₂
 v₃

julia> factor.(components(TV, (:cont,), b))
3-element Vector{Sym}:
 -(v1 - v2 - v3)/2
  (v1 - v2 + v3)/2
  (v1 + v2 - v3)/2

julia> components(TV, (:cov,), b)
3-element Vector{Sym}:
 v₂ + v₃
 v₁ + v₃
 v₁ + v₂

julia> simplify.(components(TV, (:cov,), normal_basis(b)))
3-element Vector{Sym}:
 sqrt(2)*(v2 + v3)/2
 sqrt(2)*(v1 + v3)/2
 sqrt(2)*(v1 + v2)/2

julia> T = Tensor{2,3}((i,j)->symbols("t$i$j",real=true))
3×3 Tensor{2, 3, Sym, 9}:
 t₁₁  t₁₂  t₁₃
 t₂₁  t₂₂  t₂₃
 t₃₁  t₃₂  t₃₃

julia> TT = Tensnd(T)
3×3 Tensnd{2, 3, Sym, Sym, Tensor{2, 3, Sym, 9}, CanonicalBasis{3, Sym}}:
 t₁₁  t₁₂  t₁₃
 t₂₁  t₂₂  t₂₃
 t₃₁  t₃₂  t₃₃

julia> components(TT, (:cov,:cov), b)
3×3 Matrix{Sym}:
 t₂₂ + t₂₃ + t₃₂ + t₃₃  t₂₁ + t₂₃ + t₃₁ + t₃₃  t₂₁ + t₂₂ + t₃₁ + t₃₂
 t₁₂ + t₁₃ + t₃₂ + t₃₃  t₁₁ + t₁₃ + t₃₁ + t₃₃  t₁₁ + t₁₂ + t₃₁ + t₃₂
 t₁₂ + t₁₃ + t₂₂ + t₂₃  t₁₁ + t₁₃ + t₂₁ + t₂₃  t₁₁ + t₁₂ + t₂₁ + t₂₂

julia> factor.(components(TT, (:cont,:cov), b))
3×3 Matrix{Sym}:
 -(t12 + t13 - t22 - t23 - t32 - t33)/2  …  -(t11 + t12 - t21 - t22 - t31 - t32)/2
  (t12 + t13 - t22 - t23 + t32 + t33)/2      (t11 + t12 - t21 - t22 + t31 + t32)/2
  (t12 + t13 + t22 + t23 - t32 - t33)/2      (t11 + t12 + t21 + t22 - t31 - t32)/2

TensND.fϵ — Method

fϵ(T, i::Int, j::Int, k::Int)
fϵ(i::Int, j::Int, k::Int) = fϵ(Int, i::Int, j::Int, k::Int)

Function giving Levi-Civita symbol ϵᵢⱼₖ = (i-j) (j-k) (k-i) / 2

TensND.isorthogonal — Method

isorthogonal(b::AbstractBasis)

Checks whether the basis b is orthogonal

TensND.isorthonormal — Method

isorthonormal(b::AbstractBasis)

Checks whether the basis b is orthonormal

TensND.metric — Method

metric(b::AbstractBasis, var = :cov)

Returns the covariant (if var = :cov) or contravariant (if var = :cont) metric matrix

TensND.normal_basis — Method

normal_basis(v::AbstractArray{T,2}, var = :cov) where {T}

Builds a basis after normalization of column vectors of input matrix v where columns define either

primal vectors ie eᵢ=v[:,i]/norm(v[:,i]) if var = :cov as by default

dual vector ie eⁱ=v[:,i]/norm(v[:,i]) if var = :cont.

TensND.vecbasis — Method

vecbasis(b::AbstractBasis, var = :cov)

Returns the primal (if var = :cov) or primal (if var = :cont) basis